Player 2 spawner, SO refractor, player vs player shooting

+++ Asteroid-Unity/Assets/Scripts/ScriptableObjects/SO_Bullet.cs
@@ -10,7 +10,7 @@ public class SO_Bullet : ScriptableObject
     [SerializeField] private string description;
     [SerializeField] private Sprite image;
     [SerializeField] private float speed;
-    [SerializeField] private float firePower;
+    [SerializeField] private int firePower;
     [SerializeField] private GameObject bulletLvl1;
     [SerializeField] private GameObject bulletLvl2;
     [SerializeField] private GameObject bulletLvl3;
@@ -19,9 +19,8 @@ public class SO_Bullet : ScriptableObject
     public string Description { get => description;}
     public Sprite Image { get => image;}
     public float Speed { get => speed;}
-    public float FirePower { get => firePower;}
+    public int FirePower { get => firePower;}
     public GameObject BulletLvl1 { get => bulletLvl1;}
     public GameObject BulletLvl2 { get => bulletLvl2;}
     public GameObject BulletLvl3 { get => bulletLvl3;}
-
 }
